Fix storage of device_fingerprint_reuse field as float
The `device_fingerprint_reuse` field was supposed to be an integer. The
number of times will always not have a decimal part.

But because of GRPC conversions in `result-store` we are storing this value as a float.
It is not trivial to fix this. And on openapi, a float value can be
represented as having a decimal part or not (it is optional), i.e. both
`3` and `3.0` are valid.

So it is much easier to fix this issue on the field by saying the field
is a float. Results will be valid when they are without decimal
part, but will also be valid when the api returns `0.0`.